+++ backend/clickhouse/migrations/000143_update_phonehome_tables.down.sql
@@ -1,0 +1,8 @@
+DROP TABLE IF EXISTS phonehome_mv;
+
+CREATE MATERIALIZED VIEW phonehome_mv TO phonehome AS
+select Timestamp, UUID, SpanName, ServiceName, ServiceVersion, TraceAttributes
+from traces
+where ProjectId = 1
+  and TraceAttributes['highlight.type'] = 'highlight.phonehome'
+  and TraceAttributes['highlight-doppler-config'] like 'docker%';