+++ backend/lambda-functions/deleteSessions/deleteSessionBatchFromPostgres/main.go
@@ -2,10 +2,9 @@ package main
 
 import (
 	"github.com/aws/aws-lambda-go/lambda"
-	"github.com/highlight-run/highlight/backend/env"
+	lambdafunctions "github.com/highlight-run/highlight/backend/lambda-functions"
 	"github.com/highlight-run/highlight/backend/lambda-functions/deleteSessions/handlers"
 	"github.com/highlight/highlight/sdk/highlight-go"
-	hlog "github.com/highlight/highlight/sdk/highlight-go/log"
 )
 
 var h handlers.Handlers
@@ -15,13 +14,7 @@ func init() {
 }
 
 func main() {
-	highlight.SetProjectID("1jdkoe52")
-	highlight.Start(
-		highlight.WithServiceName("lambda-functions--deleteSession-batchFromPostgres"),
-		highlight.WithServiceVersion(env.Config.Version),
-		highlight.WithEnvironment(env.EnvironmentName()),
-	)
-	hlog.Init()
+	lambdafunctions.Monitor("lambda-functions--deleteSession-batchFromPostgres")
 	lambda.StartWithOptions(
 		h.DeleteSessionBatchFromPostgres,
 		lambda.WithEnableSIGTERM(highlight.Stop),