+++ frontend/src/pages/Alerts/AlertGraph/index.tsx
@@ -1,5 +1,5 @@
 import { Box } from '@highlight-run/ui/components'
-import React, { useMemo } from 'react'
+import React from 'react'
 import { ReferenceArea, ReferenceLine } from 'recharts'
 
 import {
@@ -66,10 +66,6 @@ export const AlertGraph: React.FC<Props> = ({
 
 	const viewConfig = sessionsProduct ? BAR_CONFIG : LINE_CONFIG
 
-	const expressions = useMemo(() => {
-		return [{ aggregator: functionType, column: functionColumn }]
-	}, [functionColumn, functionType])
-
 	return (
 		<Box cssClass={style.graphWrapper} shadow="small">
 			<Box
@@ -114,7 +110,9 @@ export const AlertGraph: React.FC<Props> = ({
 						thresholdType,
 						thresholdValue,
 					}}
-					expressions={expressions}
+					expressions={[
+						{ aggregator: functionType, column: functionColumn },
+					]}
 				>
 					{!sessionsProduct &&
 						thresholdType === ThresholdType.Constant && (