+++ highlight.io/components/QuickstartContent/logging/elixir/other.tsx
@@ -9,7 +9,7 @@ export const ElixirOtherLogContent: QuickStartContent = {
 		{
 			title: 'Set up your highlight.io SDK.',
 			content: `First, make sure you've followed the [backend getting started](${siteUrl(
-				'/docs/getting-started/backend-sdk/elixir/other',
+				'/docs/getting-started/server/elixir/other',
 			)}) guide.`,
 		},
 		{