+++ highlight.io/components/QuickstartContent/server/js/shared-snippets-monitoring.tsx
@@ -76,7 +76,7 @@ H.consumeError(error, parsed?.secureSessionId, parsed?.requestId)`,
 export const setupLogging: (slug: string) => QuickStartStep = (slug) => ({
 	title: 'Set up logging.',
 	content: `With the JS SDKs, your logs are reported automatically from console methods. See the JS [logging setup guide](${siteUrl(
-		'/docs/getting-started/backend-logging/js/overview',
+		'/docs/getting-started/server/js/overview',
 	)}) for more details.`,
 })