+++ highlight.io/components/QuickstartContent/server/ruby/shared-snippets-monitoring.tsx
@@ -66,6 +66,6 @@ export const verifyErrors: QuickStartStep = {
 export const setUpLogging: (slug: string) => QuickStartStep = (slug) => ({
 	title: 'Set up logging.',
 	content: `Start sending logs to Highlight! Follow the [logging setup guide](${siteUrl(
-		`/docs/getting-started/backend-logging/ruby/${slug}`,
+		`/docs/getting-started/server/ruby/${slug}`,
 	)}) to get started.`,
 })