+++ highlight.io/components/QuickstartContent/server/rust/shared-snippets-monitoring.tsx
@@ -4,6 +4,6 @@ import { QuickStartStep } from '../../QuickstartContent'
 export const setUpLogging: (slug: string) => QuickStartStep = (slug) => ({
 	title: 'Set up logging.',
 	content: `Start sending logs to Highlight! Follow the [logging setup guide](${siteUrl(
-		`/docs/getting-started/backend-logging/rust/${slug}`,
+		`/docs/getting-started/server/rust/${slug}`,
 	)}) to get started.`,
 })